+++ src/main/java/com/server/concert_reservation/api/concert/infrastructure/repository/querydsl/ConcertScheduleCoreQueryDsl.java
@@ -3,17 +3,12 @@ package com.server.concert_reservation.api.concert.infrastructure.repository.que
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.server.concert_reservation.api.concert.infrastructure.entity.ConcertScheduleEntity;
 import com.server.concert_reservation.api.concert.infrastructure.entity.QConcertScheduleEntity;
-import com.server.concert_reservation.common.exception.CustomException;
-import jakarta.persistence.NoResultException;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
-
-import static com.server.concert_reservation.common.exception.code.ConcertErrorCode.CONCERT_SCHEDULE_NOT_FOUND;
 
 @RequiredArgsConstructor
 @Repository
