--- src/main/java/com/server/concert_reservation/api/token/common/uuid/UUIDManager.java
@@ -1,5 +1,0 @@
-package com.server.concert_reservation.api.token.common.uuid;
-
-public interface UUIDManager {
-    String generateUuid();
-}
