+++ src/main/java/com/server/concert_reservation/api/token/infrastructure/repository/core/TokenCoreReader.java
@@ -5,9 +5,9 @@ import com.server.concert_reservation.api.token.domain.repository.TokenReader;
 import com.server.concert_reservation.api.token.infrastructure.entity.TokenEntity;
 import com.server.concert_reservation.api.token.infrastructure.repository.TokenJpaRepository;
 import com.server.concert_reservation.api.token.infrastructure.repository.querydsl.TokenQueryDsl;
-import com.server.concert_reservation.common.exception.CustomException;
-import com.server.concert_reservation.common.exception.code.TokenErrorCode;
-import com.server.concert_reservation.support.time.TimeManager;
+import com.server.concert_reservation.support.api.common.exception.CustomException;
+import com.server.concert_reservation.api.token.domain.errorcode.TokenErrorCode;
+import com.server.concert_reservation.support.api.common.time.TimeManager;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
 import org.springframework.stereotype.Repository;
