+++ src/main/java/com/server/concert_reservation/support/api/common/time/TimeProvider.java
@@ -1,4 +1,4 @@
-package com.server.concert_reservation.support.time;
+package com.server.concert_reservation.support.api.common.time;
 
 import org.springframework.stereotype.Component;
