+++ src/main/java/com/server/concert_reservation/support/api/common/uuid/UUIDManager.java
@@ -1,0 +1,5 @@
+package com.server.concert_reservation.support.api.common.uuid;
+
+public interface UUIDManager {
+    String generateUuid();
+}
